+++ test/unit_problem_solving_searching.lua
@@ -30,6 +30,7 @@ local germany = require("germany")
 require("breadth_first_search")
 require("uniform_cost_search")
 require("depth_first_search")
+require("depth_limited_search")
 
 print("Breadth-first search")
 print("--------------------")
@@ -53,7 +54,7 @@ print("Elapsed: " .. t2 - t1)
 print("")
 
 print("Uniform-cost search")
-print("--------------------")
+print("-------------------")
 t1 = os.clock()
 solution = uniform_cost_search(germany.problem, 
   germany.state.Frankfurt, germany.state.Munchen, true)
@@ -76,7 +77,7 @@ print("Elapsed: " .. t2 - t1)
 print("")
 
 print("Depth-first search")
-print("--------------------")
+print("------------------")
 t1 = os.clock()
 solution = depth_first_search(germany.problem, 
   germany.state.Frankfurt, germany.state.Munchen, true)
@@ -89,6 +90,32 @@ print("Path cost is... " .. solution.cost)
 
 print("Parent 1 is Frankfurt...", 
   solution.parent[1] == germany.state.Frankfurt)
+
+print("Elapsed: " .. t2 - t1)
+
+print("")
+
+print("Depth-limited search")
+print("--------------------")
+t1 = os.clock()
+solution = depth_limited_search(germany.problem, 
+  germany.state.Frankfurt, germany.state.Munchen, 3)
+t2 = os.clock()
+
+print("Solution is Munchen...",
+  solution.state == germany.state.Munchen)
+
+print("Path cost is... " .. solution.cost)
+
+for i = 1, #solution.parent do
+  local cName = nil
+  for city,id in pairs(germany.state) do
+    if (id == solution.parent[i]) then
+      print("Parent " .. i .. ": " .. city)
+      break
+    end
+  end
+end
 
 print("Elapsed: " .. t2 - t1)